# src/main/java/egor/enrollment/components/schemas/SystemItemHistoryUnit.java
package egor.enrollment.components.schemas;

public class SystemItemHistoryUnit {
    private String id;
    private String url;
    private String date;
    private String parentId;
    private Integer size;
    private String type;
}